refactor products service from go to python
