+++ src/products/src/products-service/server.py
@@ -1,14 +1,9 @@
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 # SPDX-License-Identifier: MIT-0
-
 from flask import Flask
 from flask_cors import CORS
-import logging
 
 
 
 app = Flask(__name__)
-handler = logging.StreamHandler()
-handler.setLevel(logging.INFO)
-app.logger.addHandler(handler)
 CORS(app)
